refactor: enhance profile picture upload functionality
